refactor getUploadableFileList to share with watch command
